requirements.txt
flwr[simulation]
flwr-datasets[vision]
torch
torchvision
matplotlib
hydra-core
ray
